+++ app/Http/Livewire/Permission/Delete.php
@@ -3,10 +3,12 @@
 namespace App\Http\Livewire\Permission;
 
 use App\Models\Permission;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
     public $permission;
     public $modalDeleteVisible = false;
     public $actionShowDeleteModal = false;
@@ -19,8 +21,10 @@ class Delete extends Component
 
     public function delete($id)
     {
-
+        $this->authorize('DELETE_PERMISSION', 'PERMISSION');
         $permission = Permission::find($id);
+        cache()->forget('permissions');
+        cache()->forget('permission-' . strtoupper($permission->model));
         if ($permission) {
             $permission->delete();
         }
